adding progress (time left) and fixing a random number bug in read, update, delete

+++ drivers/cs.sh
@@ -46,7 +46,7 @@ function op_read() {
 	then
 		method="GET"
 		header_date=$(date +'%a, %d %b %Y %T %Z')
-	    line=$(($RANDOM % $filecount))
+	    line=$(($RANDOM % $filecount + 1))
 	    filename=$(awk "NR==$line" ./filelist.txt)
 	    output_filename=`(echo "$RANDOM" | md5sum | head -c 12)`
 	    path="test/$filename"
@@ -78,7 +78,7 @@ function op_update() {
 		method="PUT"
 		header_date=$(date +'%a, %d %b %Y %T %Z')
 	    filenumber=$(($RANDOM % 100))
-	    line=$(($RANDOM % $filecount))
+	    line=$(($RANDOM % $filecount + 1))
 	    filename=$(awk "NR==$line" ./filelist.txt)
 	    path="test/$filename"
 	    auth_string="$method\n\napplication/octet-stream\n$header_date\n/$path"
@@ -109,19 +109,18 @@ function op_delete() {
 	then
 		method="DELETE"
 		header_date=$(date +'%a, %d %b %Y %T %Z')
-	    filenumber=$(($RANDOM % 100))
-	    line=$(($RANDOM % $filecount))
+	    line=$(($RANDOM % $filecount + 1))
 	    filename=$(awk "NR==$line" ./filelist.txt)
-	    sed -i "$(($line))d" ./filelist.txt
 	    path="test/$filename"
 	    auth_string="$method\n\napplication/octet-stream\n$header_date\n/$path"
 	    hash_code=`echo -n -e "$auth_string" | openssl dgst -binary -sha1 -hmac $moss_secret_key | base64`
 	    auth_header="AWS $moss_access_key:$hash_code"
 	    
-	    if [ "$DEBUG" != TRUE ]
+	    if [ "$DEBUG" != TRUE ] && [ "$filename"]
 		then
 			result=`curl -o /dev/null -w "time:%{time_total},status:%{http_code}" -k -s -H "Authorization: $auth_header" -H "Content-Type: application/octet-stream" -H "Date: $header_date" -XDELETE --proxy1.0 $cs_proxy_host $cs_host/$path`
 	    	echo $result >> $results_dir/stats.txt
+	    	sed -i "$(($line))d" ./filelist.txt
 	    else
 			print_debug "Curl command:"
 			print_debug "curl -o /dev/null -w \"time:%{time_total},status:%{http_code}\" -k -s -H \"Authorization: $auth_header\" -H \"Content-Type: application/octet-stream\" -H \"Date: $header_date\" -XDELETE --proxy1.0 $cs_proxy_host $cs_host/$path"
